+++ common/src/main/java/com/human/common/gameplay/entity/living/human/marine/ai/MarineGOAPGoals.java
@@ -1,0 +1,18 @@
+package com.human.common.gameplay.entity.living.human.marine.ai;
+
+import com.just.goap.Goal;
+import com.just.goap.condition.Condition;
+import com.just.goap.condition.expression.Expression;
+
+public class MarineGOAPGoals {
+
+    public static final Goal EQUIP_BEST_ARMOR = Goal.builder(
+        "EquipBestArmorGoal",
+        new Condition<>(MarineGOAPKeys.BEST_ARMOR_SET, Expression.isNone())
+    )
+        .build();
+
+    private MarineGOAPGoals() {
+        throw new UnsupportedOperationException();
+    }
+}